Add files via upload
Modified the "Attendance" sheet and correct some rows. Some of the attendee's info for virtual events were not accurate and are corrected this time. 

There are approximately 200-400 rows with no rsvp info because these members are in the original sheets Allison gave us. They are the part of people who received the notification, did not respond but attended the event.